+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 > [!NOTE]
 > While we strive to maintain backwards compatibility as much as possible, we can't guarantee semantic versioning will be strictly followed, as this provider depends on the underlying [bunny.net API](https://docs.bunny.net/reference/bunnynet-api-overview).
 
+## [Unreleased]
+### Fixed
+- resource pullzone_edgerule: creating multiple edgerules for the same pullzone causes inconsistencies in tfstate
 
 ## [0.3.6] - 2024-08-14
 ### Fixed